Rename util to tool for fixing inconsistency with Bundler
