+++ tool/cops/deprecations.rb
