+++ tool/create_encrypted_key.rb
