+++ tool/generate_spdx_license_list.rb
